btrfs-progs: build: CPU feature checks compatible with clang
Clang, unlike GCC, typically accepts unknown or unsupported `-m`
options without error, simple ignoring them if they don't apply to
the target architecture. This means that the `AX_CHECK_COMPILE_FLAG`
macro, which relies on the compiler to fail when an unsupported
flag is used, doesn't work as intended with clang.

This change fixes that by compiling test programs which make an
actual use of tested flags. Such checks work reliably on clang.

Fixes: #712